requirements.txt
Pillow>=9.0.0
cairosvg>=2.5.0
fonttools>=4.0.0